Transfer the core functionality from aiidalab-home package. (#27)
This PR imports the core functionality of the aiidalab-home app, making
aiidalab be not just a meta-package that installs dependencies, but a
core package that sets up basic functionality of AiiDA lab.

The main modules that are transferred from aiidalab-home are:

* `app.py`: contains `AiidaLabApp` class which allows installing,
deleting and updating apps. 
* `appstore.py`: contains `AiidaLabAppStore` class which helps to
navigate the apps available in AiiDA lab registry.
* `home.py`: contains AiidaLabHome class which generates AiiDA lab
home page.
* `config.py`: contains some general settings.
* `utils.py`: contains commonly-used functions.

Also, this PR adds pre-commit hooks + Travis tests that run them.