Revise installation release line specification (#80)
Revise the specification and management of app versions and release lines.

A release line is indicated via a git reference specified as `@ref` as part of the
app's git-url. A *ref* specifies the release line, which can be a branch name, a tag, 
or a specific commit. The installed version of an app must be part of the release
line which for a branch includes all tagged commits on the branch and the tip of
the branch (if not yet included). A release line that is defined by a tag or a commit
constitutes a special case of a release line with only one available version.

An app that that has local modifications or that has a revision checked out that is not
on the specified release line is considered to be in a "detached" state, in which case
automated version management is disabled by default.

Co-authored-by: Casper Welzel Andersen <43357585+CasperWA@users.noreply.github.com>