Refactor the AiidaLabApp repository watch function into separate class.
This refactoring moves the implementation of the repository watch
function of the AiidaLabApp class into a separate class. This improves
modularity and makes it easier to manage the watches outside of the
main class.

The AiidaLabApp 'watch' constructor argument was added, which – if set
to False – prevents the automated start of the watch after object
construction. This makes this class usable in synchronous settings such
as a command line interface or other programmatic use of this class.