Refactor/aiidalab app repository watch (#85)
This refactoring moves the implementation of the repository watch
functions of the `AiidaLabApp` class into a separate class. This improves
modularity and makes it easier to manage the watches outside of the
main class.

The `AiidaLabApp` 'watch' constructor argument was added. The argument, if set
to False, prevents the automated start of the watch after object
construction. This makes this class usable in synchronous settings such
as a command-line interface or other programmatic use of this class.

Additionally, this PR improves the AiidaLabAppWatch thread control logic.