Install app dependencies with dedicated environment (#82)
Enable the management as well as automated installation of virtual
environments for each app to make it possible to install app-specific
dependencies. The environment installation process involves the creation
of a virtual environment with `venv`, the subsequent installation of the
Python executable as kernel into the Jupyter kernelspec configuration
within user space, and the installation of the app dependencies. The
virtual environment uses system packages, so dependencies that are
already present in the system environment are used directly and not
installed again.

The environment is uniquely identified by its name which is a function
of the app name to prevent accidental name collisions. To ensure that
app-specific notebooks are opened with the correct Jupyter kernel, one
must use the `aiidalab.utils.url_for()` function when creating intra-
and inter-app links. For example, to link to another notebook within the
same app, simply use `url_for('other_notebook.ipynb')` or
`url_for('other/notebook.ipynb')`. The `url_for()` function will
automatically add the correct kernel specification for a given URL. This
works for both intra- and inter-app links. If an app has no kernel
installed, this function will fallback to the default kernel
(`python3`). Links that do not use the `url_for()` function will use the
kernel specified in the notebook metadata or the default kernel, this
must therefore be avoided.

App dependencies are installed by installing the app as an editable
package if a `setup.py` file is present or – if that's not the case – by
installing dependencies from a `requirements.txt` file if present. If
neither file is present, no environment and no dependencies are
installed.

Co-authored-by: Aliaksandr Yakutovich <yakutovicha@gmail.com>