Require post_install script to be an executable.
Instead of executing it with /bin/bash by default, we run the script as
an executable.

There is no reason for requiring the post_install script to be a bash
script, it can be any executable. App developers can easily create an
executable bash script by adding the `#!/bin/bash` shebang line at the
top of the script.