Use the number sign '#' instead of '@' to indicate release line.
This is more in line with the URL specification (RFC 3986, 3.5) that
defines that the fragment section of a URL is not transmitted to the
server endpoint and evaluated on the client side to denote "some portion
or subset of the primary resource".

Using '#' instead of '@' also avoids potential confusion with the use of
the symbol in the context of `pip install`, which has highly related,
but not identical meaning.