Merge stdour and stderr for easier handling.
The issue is that both process.stdout.readline() and
process.stderr.readline() are blocking the execution till the
first message pops up. In case no errors/warnings printed, one
needs to wait, essentially, till the end of the scrip exectuion
for the log to appear. To overcome this problem we merge STDOUT
and STDERR.