docs: note for how to develop when UID of host and container different
A note is added to app create Step 5, which describe how to add code to
make the app functional. When running aiidalab locally, the UID of
aiidalab's bind mounted home folder may different from the UID of host
user. This lead to editing files inside folder permission denied because
of the ownership issue.
Vscode has a feature "Attach to a running container" to make
development inside the container possible.