add count to proper set busy traitlets (#357)
This is the more direct cause of aiidalab/aiidalab-home#100
In the package, there is already a context manager _show_busy provide to hold the display before the operations are running. However, this context manager only set the `self.busy` to a bool, which will be overridden unexpectedly if the _show_busy is embedded in itself. This happened for install_app, updata_app and uninstall_app, they are supposed to use _show_busy to prevent information to be displayed, but inside these function also call refresh which have another _show_busy. I add a counter to properly set the busy traitlets. It is False (which means not busy) when it counts down to zero.