validate the core dependencies and show packages to install for new version (#347)
After migrating the docker stack to support AiiDA 2.x, most apps in the appstore would incompatible since they are specifically set to dependent on aiida-core~=1.0. We introduce in this PR the concept of "core packages", it is a list of dependencies that the app must fullfill in the requirement dependencies list. The core packages defined at the moment are `aiida-core`, `jupyter-client`, we can bring more in the future such as aiidalab itself and the aiidalab-home, which we don't want the app to override. These packages are in principle installed in advance in docker stack and shipped with the aiidalab docker image.

In this commit, we also find and correct the long stand issue that some spurious incompatible popped for the apps, such as `Vapory` and `jupyter-client`. The causing is from the package name is not always match with the name used in the requirement list, for some packages, the package name in the libraries has an underscore as the connector (un-canonical name) but in the requirement list of the app, it has the name with the dash connected (canonical name).  The package name with letters in uppercase is also non-canonical e.g `Vapory`.

Meanwhile, we bring the unit test and mock the AiidalabAppManager dataclass to run the unit test, although not fill all the test gap but already improve the robustness of the touched code.

---------

Co-authored-by: Aliaksandr Yakutovich <yakutovicha@gmail.com>
Co-authored-by: Daniel Hollas <danekhollas@gmail.com>