Catch all exceptions when installing app
Previously, we were only catching RuntimeError which
left a lot of room for other failures during installation,
which would result in an incomplete installation of an app.
Generally, there's no way of knowing after-the-fact that
the installation ended prematurely,
e.g. `aiidalab list` would show the app as installed.

Note that this includes for example post_install script failures.

Now we catch everything, including Keyboard interrupts,
so that we always trigger the already existing rollback mechanism.

See issue #334 for details of how to reproduce this.

Test plan:
1. Trigger installation of an app from CLI
   `aiidalab -v install aurora==0.2.1`
2. Interrupt the installation in the middle with Ctrl+C
3. The installation should rollback. Verify with `aiidalab list`
   that the app is not installed.
4. Try installing an app from a following URL, where I specifically
   modified the `post_install` script to fail.
   `aiidalab -v install aiidalab-ispg@git+https://github.com/ispg-group/aiidalab-ispg@post-install-fail`
5. Failing post_install should now trigger rollback as well.

Closes #334