Using event to log the stop status of the monitor threading (#367)
With using mypy to check the function types and calling issues, we found that in the `AiidalabAppWatch` class, the thread creates and uses the attribute `_stop_flag` dynamically which was better to be replaced by using the thread event to record the stop status of the thread. 
The unit test was also added to catch the errors in the future.