available versions list unified by consider prereleases and core compatible in same function (#368)
fixes #360

The available_versions of _AiidalabApp data class object not considered core compatible. The core compatible check is implement independently and sporadically when the available_versions list is required.
This commit add the core compatible check inside the function available_versions of _AiidalabApp class to get the only usable versions for both when prereleases option is on and off. Therefore the version list conforms with what is shown in the version_to_install dropdown option.