incompatibility check for app detached will go through the local repo (#374)
fixes #373

When the app is installed but the app repo gets a new commit, the app version will become UNKNOWN. Therefore, the requirements list cannot check with the version from the registered list. Rather the requirements list needs to access from the local repo.
Check requirements upon local repo happened for the app not registered but also should be for the app detached. The condition is added for the detached app for incompatibility check.