Fix base.html URLs
Updated the URLs used in the footer of the application to use the correct links. Previously, the links were pointing to the wrong resources, but have been updated to point to the correct URLs. This includes updating the links to the AiiDAlab registry page in the aiidalab-registry repository. These changes ensure that users can access the correct resources from the footer of the application.