support using * for matching all branches in getting registry (#383)
The release_line of app registry has to be the a rev selection list from a certain branch, this makes it impossible to have the tags collected from other branches. This PR makes it possible by supporting the setting branch using wildcard character * to search the rev from all the branches.