Skip invalid app versions (#451)
The AiiDAlab App versions relies on Git tags. 
However, we have an implicit assumption that the tags are a semantically valid versions,
 specifically versions valid according to PEP440. 
Since we do not have control over the apps, this cannot be enforced in any way.
New versions of `packaging.parse()` function that we use
all over the place started to throw `InvalidVersion` exception
when given an invalid version, which lead to crashes
described in #339 and aiidalab/aiidalab-registry#103.

I went over the all instances of packaging.parse
and made sure the call sites are safe. Finally, this allows us
to unpin the `packaging` dependency.

* Skip invalid versions in _AiidaLabApp.releases
* Show test durations
* Add unit tests for sort_semantic, add 'reverse' parameter